Only interact with riff k8s resource definitions
Some `riff` commands attempt to interact with all kubernaties resources
defined within a directory. This can fail if that directory contains
other `*.json` or `*.yaml` files.

The `riff` cli should only attempt to interact with files that match
`*-function.yaml` or `*-topics.yaml` in order to avoid potential
collisions with other k8s resource definitions or files that are not
resource definitions but have the same file extensions.